[frames] split into frames header and frames-deprecated
